chore: convert index.test.js to ts
